[server] | refactor(): 상대 경로로 수정
